+++ test/types/models.test.ts
@@ -9,6 +9,7 @@ import mongoose, {
   CallbackError,
   HydratedDocument,
   HydratedDocumentFromSchema,
+  InsertManyResult,
   Query,
   UpdateWriteOpResult,
   AggregateOptions,
@@ -708,4 +709,45 @@ async function gh13746() {
   expectType<boolean | undefined>(findOneAndUpdateRes.lastErrorObject?.updatedExisting);
   expectType<ObjectId | undefined>(findOneAndUpdateRes.lastErrorObject?.upserted);
   expectType<OkType>(findOneAndUpdateRes.ok);
+}
+
+function gh13904() {
+  const schema = new Schema({ name: String });
+
+  interface ITest {
+    name?: string;
+  }
+  const Test = model<ITest>('Test', schema);
+
+  expectAssignable<Promise<InsertManyResult<ITest>>>(Test.insertMany(
+    [{ name: 'test' }],
+    {
+      ordered: false,
+      rawResult: true
+    }
+  ));
+}
+
+function gh13957() {
+  class RepositoryBase<T> {
+    protected model: mongoose.Model<T>;
+
+    constructor(schemaModel: mongoose.Model<T>) {
+      this.model = schemaModel;
+    }
+
+    // Testing that the following compiles successfully
+    async insertMany(elems: T[]): Promise<T[]> {
+      elems = await this.model.insertMany(elems);
+      return elems;
+    }
+  }
+
+  interface ITest {
+    name?: string
+  }
+  const schema = new Schema({ name: String });
+  const TestModel = model('Test', schema);
+  const repository = new RepositoryBase<ITest>(TestModel);
+  expectType<Promise<ITest[]>>(repository.insertMany([{ name: 'test' }]));
 }